scaleway_compute: add "terminate" option to use with "absent" state
Using "absent" state destroys the server but keeps its volumes unattached.
To completely destroy the server with volumes the Scaleway API has
undocumented yet but provided "terminate" action instead of "poweroff".
This option is only applicable to running servers, and there is no
equivalent API command to completely destroy stopped servers.

This patch adds "terminate: yes/no" module option. If new state is
"absent" and "terminate" is set to yes, it will use "terminate" action
to stop and destroy server.

If server is already stopped, this will not destroy it and return error.
It is not an expected behaviour, but this way the Scaleway API works.
It can be worked around by using same module twice with terminate and
without it to destroy server and keeping volumes at worst. But better
would be if Scaleway allows this action for stopped servers.